+++ tests/load_testdata.R
@@ -3,6 +3,8 @@ test_docs <- readRDS(system.file("testdata", "idaifield_test_docs.RDS",
                                  package = "idaifieldR"))
 test_resources <- check_and_unnest(test_docs)
 
+test_simple <- simplify_idaifield(test_resources)
+
 config <- attributes(test_docs)$config
 
 
@@ -12,12 +14,12 @@ skip_if_no_connection <- function() {
                       pwd = "hallo", ping = FALSE)
     )
 
-  ping <- try(idf_ping(connection))
+  ping <- suppressWarnings(idf_ping(connection))
 
-  if (!inherits(ping, "try-error")) {
+  if (ping) {
     connection$status <- idf_ping(connection)
     return(connection)
+  } else {
+    skip("Test skipped, needs DB-connection")
   }
-
-  skip("Test skipped, needs DB-connection")
 }