+++ tests/testthat/test-check_if_idaifield.R
@@ -1,27 +1,22 @@
 source(file = "../load_testdata.R")
 
 test_that("checking works for docs-lists", {
-  check <- check_if_idaifield(test_docs)
+  docs <- list(1, 2, 3)
+  class(docs) <- "idaifield_docs"
+  check <- check_if_idaifield(docs)
   expect_true(check["idaifield_docs"])
 })
 
-
-test_that("checking works for docs-lists", {
-  check <- check_if_idaifield(test_docs)
-  expect_false(check["idaifield_resources"])
-})
-
 test_that("checking works for resource-lists", {
-  check <- check_if_idaifield(test_resources)
-  expect_true(check["idaifield_resources"])
-})
-
-test_that("checking works for resource-lists", {
-  check <- check_if_idaifield(unnest_docs(test_docs))
+  resources <- list(1, 2, 3)
+  class(resources) <- "idaifield_resources"
+  check <- check_if_idaifield(resources)
   expect_true(check["idaifield_resources"])
 })
 
 test_that("checking works for simplified lists", {
-  check <- check_if_idaifield(simplify_idaifield(test_docs))
+  simple <- list(1, 2, 3)
+  class(simple) <- "idaifield_simple"
+  check <- check_if_idaifield(simple)
   expect_true(check["idaifield_simple"])
 })