+++ NEWS.md
@@ -1,4 +1,9 @@
-# idaifieldR 0.3.1 _tba_
+# idaifieldR 0.3.2 _tba_
+* add option to use or not use exact dates as min/max values for dating if present
+* remove config as attribute, that was a bad idea
+* rename `select_by()` to `idf_select_by()`
+* rename `show_categories()` to `idf_show_categories()`
+* update `idf_select_by()` and `idf_show_categories()`
 
 
 # idaifieldR 0.3.1 _2023-04-02_